- Performance improvements:   - Switched from traditional switch statement to expression-based switch for more efficient dispatch   - Delayed cell object allocation until after record type validation, avoiding unnecessary object creation for invalid records   - Use of `[MethodImpl(MethodImplOptions.AggressiveInlining)]`

+++ Excel_PRIME/XlsbImp/PooledRecordBuffer.cs
@@ -43,7 +43,7 @@ internal sealed class PooledRecordBuffer : IDisposable
 
     public short GetInt16(int offset) => BitConverter.ToInt16(_array, offset);
 
-    public short GetByte(int offset) => _array[offset];
+    public byte GetByte(int offset) => _array[offset];
 
     public string GetString(int offset)
     {
